% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{+.ssmulti}
\alias{+.ssmulti}
\title{ssmulti Addition}
\usage{
\method{+}{ssmulti}(a, b)
}
\arguments{
\item{a}{ssmulti object.}

\item{b}{ssmulti object.}
}
\description{
Combine two multi component supplystack objects
}

