% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{+.supplystack}
\alias{+.supplystack}
\title{supplystack Addition}
\usage{
\method{+}{supplystack}(a, b)
}
\arguments{
\item{a}{supplystack object.}

\item{b}{supplystack object.}
}
\description{
Combine two supplystack objects
}
\examples{
supplystack(p=c(100, 50), q=rep(50, 2)) + supplystack(p=c(75, 150), q=rep(20, 2))
}

