% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ssmulti}
\alias{ssmulti}
\title{ssmulti Constructor}
\usage{
ssmulti(p, q, nms = NULL)
}
\arguments{
\item{p}{Numeric matrix. Represents cost or price for each producer. Column values belong to a single producer; rows represent different cost components.}

\item{q}{Numeric vector. Represents quantity for each producer.}

\item{nms}{Character vector (optional). Producer names for visualization.}
}
\description{
Create a supplystack object with multiple cost components
}

