% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{supplystack}
\alias{supplystack}
\title{supplystack Constructor}
\usage{
supplystack(p, q, nms = NULL)
}
\arguments{
\item{p}{Numeric vector. Represents cost or price for each producer.}

\item{q}{Numeric vector. Represents quantity for each producer.}

\item{nms}{Character vector (optional). Producer names for visualization.}
}
\description{
Create a supplystack object
}
\examples{
supplystack(p=c(100, 50, 75), q=rep(50, 3), nms=c('A','B','C'))
}

