% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.supplystack}
\alias{plot.supplystack}
\title{plot.supplystack}
\usage{
\method{plot}{supplystack}(x, plot_names = TRUE, names_cutoff = 0.1,
  rect_col = "lightblue", rect_border = "white", ...)
}
\arguments{
\item{x}{supplystack object to be plotted.}

\item{plot_names}{Optional boolean. Should names be plotted on the stack?}

\item{names_cutoff}{Optional numeric. Cutoff for plotting of names. Suppliers with total Q less than this proportion of the total will not be plotted. Defaults to 0.10.}

\item{rect_col}{Optional character. Color of shaded rectangles or NULL for none.}

\item{rect_border}{Optional character. Color of rectangle borders or NA for none.}

\item{...}{Other graphical arguments excluding "ann".}
}
\description{
Plot a supplystack object
}
\examples{
plot(supplystack(c(10, 20, 30), c(30, 20, 10)), rect_border="darkblue")
}

